Organize strategy tests folder (#18377)
