+++ tests/tests_pytorch/strategies/test_custom_strategy.py
