+++ tests/tests_pytorch/strategies/test_single_device.py
